+++ CHANGELOG.md
@@ -9,7 +9,10 @@ All Notable changes to `Csv` will be documented in this file
 - `League\Csv\BOM`
 - `League\Csv\Statement`
 - `League\Csv\RecordSet`
-- `League\Csv\XMLEncoder`
+- `League\Csv\Converter`
+- `League\Csv\XMLConverter`
+- `League\Csv\HTMLConverter`
+- `League\Csv\JsonConverter`
 - `League\Csv\is_bom` function
 - `League\Csv\bom_match` function
 - `League\Csv\Exception\CsvException`